Security fix

+++ tests/providers/google/cloud/operators/test_gcs_system_helper.py
@@ -25,17 +25,17 @@ from airflow.providers.google.cloud.example_dags.example_gcs import (
     PATH_TO_TRANSFORM_SCRIPT,
     PATH_TO_UPLOAD_FILE,
 )
-from tests.test_utils.logging_command_executor import CommandExecutor
+from tests.test_utils.logging_command_executor import LoggingCommandExecutor
 
 
-class GcsSystemTestHelper(CommandExecutor):
+class GcsSystemTestHelper(LoggingCommandExecutor):
     @staticmethod
-    def create_file_to_upload():
+    def create_test_file():
+        # Create test file for upload
         with open(PATH_TO_UPLOAD_FILE, "w+") as file:
             file.writelines(["This is a test file"])
 
-    @staticmethod
-    def create_script_to_transform():
+        # Create script for transform operator
         with open(PATH_TO_TRANSFORM_SCRIPT, "w+") as file:
             file.write(
                 """import sys
@@ -51,16 +51,13 @@ with open(source, "r") as src, open(destination, "w+") as dest:
             )
 
     @staticmethod
-    def remove_file_to_upload():
-        os.remove(PATH_TO_UPLOAD_FILE)
-
-    @staticmethod
-    def remove_script_to_transform():
-        os.remove(PATH_TO_TRANSFORM_SCRIPT)
-
-    @staticmethod
-    def remove_saved_file():
-        os.remove(PATH_TO_SAVED_FILE)
+    def remove_test_files():
+        if os.path.exists(PATH_TO_UPLOAD_FILE):
+            os.remove(PATH_TO_UPLOAD_FILE)
+        if os.path.exists(PATH_TO_SAVED_FILE):
+            os.remove(PATH_TO_SAVED_FILE)
+        if os.path.exists(PATH_TO_TRANSFORM_SCRIPT):
+            os.remove(PATH_TO_TRANSFORM_SCRIPT)
 
     def remove_bucket(self):
         self.execute_cmd(["gsutil", "rm", "-r", f"gs://{BUCKET_1}"])
